refac(rust): move txt file handling logic into own file and loadout modification as well
